maded add a participantcompetition with Group and auto choose categories

+++ Models/Interfaces/IWeightCategoryRepository.cs
@@ -9,6 +9,6 @@ namespace DoJudo.Models.Interfaces
 {
     internal interface IWeightCategoryRepository : IBaseRepository<WeightCategory>
     {
-        WeightCategory GetWeightCategory(float weight);
+        WeightCategory GetWeightCategory(double weight);
     }
 }
